+++ src/Rendering/OpenGL/CubismRenderer_OpenGLES2.cpp
@@ -1745,6 +1745,11 @@ CubismRenderer_OpenGLES2::CubismRenderer_OpenGLES2() : _clippingManager(NULL)
 CubismRenderer_OpenGLES2::~CubismRenderer_OpenGLES2()
 {
     CSM_DELETE_SELF(CubismClippingManager_OpenGLES2, _clippingManager);
+
+	if (_offscreenFrameBuffer.IsValid())
+	{
+		_offscreenFrameBuffer.DestroyOffscreenFrame();
+	}
 }
 
 void CubismRenderer_OpenGLES2::DoStaticRelease()